requirements.txt
MouseInfo==0.1.3
numpy==2.2.1
pandas==2.2.3
pillow==11.1.0
poetry==1.8.5
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
six==1.17.0
tzdata==2024.2
websockets==14.1
